requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Security & Auth
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.0

# Database drivers
aiosqlite==0.19.0

# Content Generation & NLP
jinja2==3.1.3
markdown==3.5.2
beautifulsoup4==4.12.3
langdetect==1.0.9
nltk==3.8.1

# HTTP & APIs
httpx==0.26.0
aiohttp==3.9.1

# Search & Vectors
chromadb==0.4.22
sentence-transformers==2.3.1

# Task Queue & Cache
redis==5.0.1
rq==1.16.1

# Utilities
python-slugify==8.0.2
pytz==2024.1
pendulum==3.0.0
typer==0.9.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
bandit==1.7.6
safety==3.0.1

# Dev Tools
ipython==8.20.0
